+++ lib/webparts/itAssetManager/components/Dashboard.d.ts
@@ -1,9 +1,7 @@
 import * as React from 'react';
-import { AssetService } from '../services/AssetService';
 import { IAsset } from '../models/IAsset';
 interface IDashboardProps {
     assets: IAsset[];
-    assetService: AssetService;
     onViewAll: () => void;
     onViewAsset: (asset: IAsset) => void;
 }